requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==38.0.1
Django==4.1.1
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
pycparser==2.21
pytz==2022.2.1
sqlparse==0.4.2
